refactor(hardware): Fix up typing after numpy bump
Bumping numpy revealed some real issues with our code and also some real
issues with numpy's as-yet incomplete typing, such as
- many functions particularly in numpy.linalg aren't typed, like `norm`
- many arithmetic ops and functions behave quite strangely, often
returning a `np.Float[_64bit, Any]` instead of a float64 equivalent
- ndarray typing is still not quite there, but that's mostly because
some typing peps need to be merged first

This should be a low risk set of changes.

+++ hardware/tests/opentrons_hardware/hardware_control/test_move_group_runner.py
@@ -1,6 +1,7 @@
 """Tests for the move scheduler."""
 import pytest
 from typing import List, Tuple, Any
+from numpy import float64
 from mock import AsyncMock, call, MagicMock
 from opentrons_hardware.firmware_bindings import ArbitrationId, ArbitrationIdParts
 
@@ -55,7 +56,9 @@ def move_group_single() -> MoveGroups:
         [
             {
                 NodeId.head: MoveGroupSingleAxisStep(
-                    distance_mm=246, velocity_mm_sec=2, duration_sec=123
+                    distance_mm=float64(246),
+                    velocity_mm_sec=float64(2),
+                    duration_sec=float64(123),
                 )
             }
         ]
@@ -70,10 +73,10 @@ def move_group_home_single() -> MoveGroups:
         [
             {
                 NodeId.head: MoveGroupSingleAxisStep(
-                    distance_mm=0,
-                    velocity_mm_sec=235,
-                    duration_sec=2142,
-                    acceleration_mm_sec_sq=1000,
+                    distance_mm=float64(0),
+                    velocity_mm_sec=float64(235),
+                    duration_sec=float64(2142),
+                    acceleration_mm_sec_sq=float64(1000),
                     stop_condition=MoveStopCondition.limit_switch,
                     move_type=MoveType.home,
                 ),
@@ -90,10 +93,10 @@ def move_group_multiple() -> MoveGroups:
         [
             {
                 NodeId.head: MoveGroupSingleAxisStep(
-                    distance_mm=229,
-                    velocity_mm_sec=235,
-                    duration_sec=2142,
-                    acceleration_mm_sec_sq=1000,
+                    distance_mm=float64(229),
+                    velocity_mm_sec=float64(235),
+                    duration_sec=float64(2142),
+                    acceleration_mm_sec_sq=float64(1000),
                 ),
             }
         ],
@@ -101,16 +104,16 @@ def move_group_multiple() -> MoveGroups:
         [
             {
                 NodeId.gantry_x: MoveGroupSingleAxisStep(
-                    distance_mm=522,
-                    velocity_mm_sec=22,
-                    duration_sec=1,
-                    acceleration_mm_sec_sq=1000,
+                    distance_mm=float64(522),
+                    velocity_mm_sec=float64(22),
+                    duration_sec=float64(1),
+                    acceleration_mm_sec_sq=float64(1000),
                 ),
                 NodeId.gantry_y: MoveGroupSingleAxisStep(
-                    distance_mm=25,
-                    velocity_mm_sec=23,
-                    duration_sec=0,
-                    acceleration_mm_sec_sq=1000,
+                    distance_mm=float64(25),
+                    velocity_mm_sec=float64(23),
+                    duration_sec=float64(0),
+                    acceleration_mm_sec_sq=float64(1000),
                 ),
             }
         ],
@@ -118,18 +121,18 @@ def move_group_multiple() -> MoveGroups:
         [
             {
                 NodeId.pipette_left: MoveGroupSingleAxisStep(
-                    distance_mm=12,
-                    velocity_mm_sec=-23,
-                    duration_sec=1234,
-                    acceleration_mm_sec_sq=1000,
+                    distance_mm=float64(12),
+                    velocity_mm_sec=float64(-23),
+                    duration_sec=float64(1234),
+                    acceleration_mm_sec_sq=float64(1000),
                 ),
             },
             {
                 NodeId.pipette_left: MoveGroupSingleAxisStep(
-                    distance_mm=12,
-                    velocity_mm_sec=23,
-                    duration_sec=1234,
-                    acceleration_mm_sec_sq=1000,
+                    distance_mm=float64(12),
+                    velocity_mm_sec=float64(23),
+                    duration_sec=float64(1234),
+                    acceleration_mm_sec_sq=float64(1000),
                 ),
             },
         ],
